Adapt comment style to the other scripts
No comments in the source files we create, minimal comments in the install script.

This makes it easier to spot which comments are neccessary. So far, Symfony
is the only outlier here as it parses the comment to use it for the routing.
Having no explanatory comments in the source files makes it easy to spot
this.